MOVE: Manifest.ts
