+++ manifest.ts
@@ -1,7 +1,6 @@
-import packageJson from "../package.json";
-import { ManifestType } from "@src/manifest-type";
+import packageJson from "./package.json";
 
-const manifest: ManifestType = {
+const manifest: chrome.runtime.ManifestV3 = {
   manifest_version: 3,
   name: packageJson.name,
   version: packageJson.version,
